LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/template-shadowrootmode-script-src.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""


FAIL Speculative parsing, page load: template-shadowrootmode-script-src Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
